paramconfig:1)uipage support escape; 2)param modified notify uipage

+++ qtlearnings/paramconfig/UiCommon/cenablemngr.cpp
@@ -11,13 +11,13 @@ CEnableMngr* CEnableMngr::instance()
 void CEnableMngr::registerEableUi(CKeyDnComboBox* pCmb, uint16_t* pVal, uint16_t val, QWidget* w)
 {
     if (nullptr != pCmb && nullptr != pVal && nullptr != w){
-        connect(pCmb, SIGNAL(sig_valueChanged(uint16_t*, uint16_t)), this, SLOT(slot_valueChanged(uint16_t*, uint16_t)));
+        connect(pCmb, SIGNAL(sig_valueChanged(uint16_t*, uint32_t)), this, SLOT(slot_valueChanged(uint16_t*, uint32_t)));
         valUiMap[pVal][val].push_back(w);
 
         slot_valueChanged(pVal, *pVal);//set init state
     }
 }
-void CEnableMngr::slot_valueChanged(uint16_t* pVal, uint16_t valNew)
+void CEnableMngr::slot_valueChanged(uint16_t* pVal, uint32_t valNew)
 {
     qDebug()<<"addr "<<pVal<<" value "<<valNew<<" value "<<*pVal;
     auto itOut = valUiMap.find(pVal);
